+++ api/endpoints/version/routes.py
@@ -3,13 +3,13 @@ from flask_jwt_extended import create_access_token
 import requests
 
 version_endpoint = Blueprint('version', __name__)
-current_version = "3.3.1"
+current_version = "3.5.1"
 
 
 @version_endpoint.route('/api/version', methods=['GET'])
 def version():
     try:
-        response = requests.get("https://git.nilu.no/api/v4/projects/983/repository/tags")
+        response = requests.get("https://api.github.com/repos/eea/raven-administration/tags")
         response.raise_for_status()
         j = response.json()
         return {"current": current_version, "latest": j[0]["name"].replace("v.", "")}