+++ client/src/main.js
@@ -1,7 +1,8 @@
-import { createApp } from 'vue'
-import App from './App.vue'
+import { createApp } from "vue";
+import App from "./App.vue";
 import "./assets/tailwind.css";
 import "./assets/n-elements.css";
+import { disableTextSelectOnShiftDown } from "./helpers/utils";
 
 // INTERCEPT
 import intercetor from "./helpers/interceptor";
@@ -9,10 +10,22 @@ intercetor.request();
 intercetor.response();
 intercetor.default();
 
-const app = createApp(App); 
+const app = createApp(App);
+
+// TOOLTIP
+import Tooltip from "vue-follow-tooltip";
+app.use(Tooltip, {
+  delay: 100,
+  center: true,
+  offsetX: 0,
+  offsetY: -60
+});
+
+// Disable text selection on shift down
+disableTextSelectOnShiftDown();
 
 // ROUTER
-import router from "./router"
+import router from "./router";
 app.use(router);
 
-app.mount('#app')
+app.mount("#app");